백준 조건문 2번까지 해봄

+++ src/Coding01/Quiz09.java
@@ -6,11 +6,11 @@ public class Quiz09 {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
 
-        int num1 = sc.nextInt(), num2 = sc.nextInt();
-        System.out.println(num1+num2);
-        System.out.println(num1-num2);
-        System.out.println(num1*num2);
-        System.out.println(num1/num2);
-        System.out.println(num1%num2);
+        double num1 = sc.nextDouble(), num2 = sc.nextDouble();
+        System.out.println((int) (num1 + num2));
+        System.out.println((int) (num1 - num2));
+        System.out.println((int) (num1 * num2));
+        System.out.println((int) (num1 / num2));
+        System.out.println((int) (num1 % num2));
     }
 }
